news-app-backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-decouple>=3.8
Pillow>=10.0.0
requests>=2.31.0
